Use .hasSize() instead of .size()).isEqualTo()
Additionaly, minor code improvements and warning fixes.

+++ server/src/test/java/org/elasticsearch/common/bytes/PagedBytesReferenceTests.java
@@ -45,7 +45,7 @@ public class PagedBytesReferenceTests extends AbstractBytesReferenceTestCase {
         for (int i = 0; i < length; i++) {
             byteArray.set(i, (byte) random().nextInt(1 << 8));
         }
-        assertThat(byteArray.size()).isEqualTo((long) length);
+        assertThat(byteArray.size()).isEqualTo(length);
         BytesReference ref = BytesReference.fromByteArray(byteArray, length);
         assertThat(ref.length()).isEqualTo(length);
         if (byteArray.hasArray()) {
